Fixed genFakeICs, added real idealICs.
genFakeICs was filling up pixels outside
of the circle with NaN. This lead to the interpolation
function of applyInterpolatedICs to crash. Fixed this
by given the function a keyword argument fill=0.

Added true ideal ICs to simulation and added appropiate
attributes to simulation class and GUI.

Fixed problems with pinning when there was no data given.

Added more docstrings.