Added geometry element extraction.
To extract surfaces that need to be refined as a boundary layer,
we needed a way to find all elements that are necessary to construct
this surface. getSubElements and getAllSubElements fixes this.

Moreover, to be able to arbitrarly insert elements into a domain,
we needed a automatic way to do so. Added an option to do so into insertElement
via getTypeList and getTypeListName.

gmshElement.extract allows to extract any gmshElement into a new domain.

TODO:
	- ROI methods to generate attractors/boundary layers around.
	- Embryo methods to generate attractors/boundary layers around ROIs for easy/fast setup.
	- Multiple test scripts
	- Might need a setup.py --openscad option
	- Gmsh seems to have problems with complicated surfaces for boundary layer meshes. Need to find a way to make that faster.
	- Move mesh dependent vtk functions to pyfrp_vtk_module.
	- 3D simulation plotting via vtk.
	- Test boundary layer mesh generation, create tests etc.

Additions:
	- Added gmshElement.getTypeListName
	- Added gmshElement.getTypeList
	- Added getSubElements to each kind of gmshElement element
	- Added getAllSubElements and respective unittest
	- Added gmshElement.extract

Changes:
	- pyfrp_gmsh_geometry.domain.insertElement now has an autodetect mode.

Bugfixes: