Modified copyMeshFiles to also work for merged geo
So far, copyMeshFiles only copied the final .geo and .msh file. However,
when the geo file is a merged file and depends on different other .msh/.geo files,
it is necessary to move all used files to the new location such that
the mesh can be modified/regenerated if necessary. Now the function
takes care of this, updates the merged geo file and also uses a better
copy function, making performance better cross-platform.

TODO:
	- Embryo methods to generate attractors/boundary layers around ROIs for easy/fast setup.
	- Multiple test scripts
	- Might need a setup.py --openscad option
	- Gmsh seems to have problems with complicated surfaces for boundary layer meshes. Need to find a way to make that faster.
	- Move mesh dependent vtk functions to pyfrp_vtk_module.
	- 3D simulation plotting via vtk.

Additions:

Changes:
	- pyfrp_IO_module.copyMeshFiles now works for merged files.

Bugfixes: