applyInterpolatedICs fills before interpolation
So far, when we were calling applyInterpolatedICs, we first interpolated the first image of
the dataset and then applied the rim concentration to outside nodes. However, suppose
mesh nodes close to the masterROI boundary where outside of the boundary we have steep
gradient to a much smaller value than inside the masterROI. This leads to a steep interpolation
and thus too a probably small value for the node close to the boundary. To avoid this, we
now have the option to fill everything outside of the masterROI in the ICimg with concRim
before the interpolation, avoiding such artefacts.

Moreover, mapOntoImgs had a similar problem. To avoid this, we added the option to
computeInterpolatedSolutionToImg to choose different interpolation methods and then
extended mapOntoImgs to fill all values outside of ROI (if specified) with a fillValue.
Also, scaling images for output seems to also skew actual simulation data sometimes,
so mapOntoImgs also now allows not to scale images.

Print messages for PyFRAP errors/warnings/notes now also print the name of the function
they were called in. This improves backtracing.

Contour plots are sometimes hard to read, especially when investigating initial conditions.
Hence plotSolutionVariable, showIC and showICimg now allow also to plot as 3D surface plots.

TODO:
	- Embryo methods to generate attractors/boundary layers around ROIs for easy/fast setup.
	- Multiple test scripts
	- Might need a setup.py --openscad option
	- Gmsh seems to have problems with complicated surfaces for boundary layer meshes. Need to find a way to make that faster.
	- Move mesh dependent vtk functions to pyfrp_vtk_module.
	- 3D simulation plotting via vtk.
	- We might want to move away from saving to pickle files. Probably YAML, Json? However, might need to write my own parser for that.

Additions:
	- Added pyfrp_term_module.getFunctionCall.

Changes:
	- Now printWarning can also print the function name that called it.
	- Now printNote can also print the function name that called it.
	- Now printError can also print the function name that called it.
	- pyfrp_plot_module.plotSolutionVariable now can also plot as surface plot.
	- pyfrp_ROI.plotSolutionVariable now can also plot as surface plot.
	- pyfrp_simulation.showIC now can also plot as surface plot.
	- pyfrp_simulation.showICimg now can also plot as surface plot.
	- pyfrp_mesh.plotCellCenters now can be restricted to ROI
	- pyfrp_simulation.computeInterpolatedSolutionToImg now takes different interpolation methods as argument.
	- pyfrp_simulation.mapOntoImgs now takes different interpolation methods as argument and can fill everything outside of ROI with fill value.
	- pyfrp_sim_module.applyInterpolatedICs now fills the IC image outside of the masterROI with concRim.

Bugfixes:
	- Fixed bug in pyfrp_img_module.saveImg: Now scale argument actually does sth.