Temporarily PATH export added/install flag change
Added examples/scripting/simpleAnalysis.py to illustrate
how PyFRAP can be easily scripted.

Changed command line arguments of PyFRAP. Now, by default,
PyFRAP automatically downloads Gmsh. The --noGmsh flag
will prevent that. This is necessary such that we can run
``pip install pyfrp`` without passing the --gmsh flag to
all required packages and getting pip to crash.

In newer numpy versions, comparing ``numpy.ndarrays`` with None
will result in ``ValueError``. We thus now need to substitute
these checks with ``isinstance(x,np.ndarray)``. Modified
``pyfrp_simulation.simulation.run`` such that this is conform.

FiPy needs the gmsh executable to be in the PATH environment variable
of the current session. This is a pain for users when they are installing
gmsh/pyfrap. A workaround is to add gmsh to the environment's PATH variable
before execution. This also allows PyFRAP to be executed in virtualenv.
Therefore we added ``addToPATHs`` and ``addGmshToPATHs``. Both are executed
whenever we call ``runGmsh``.

TODO:

	- Data structure:
		- We might want to move away from saving to pickle files. IDEA:
			- molecule files do not save anymore list of embryos but just list of path to embryo files.
			- whenever embryo is needed, molecule calls getEmbryo and thus loading file. This also happens
			  user clicks on gui on embryo
			- Could also include a loadAll option.
	- GUI integration:
		- mesh merging
		- Simulation visualization
		- Geometry visualization
		- Multiple plotting functions
		- Embryo slicing
		- Simulation to image export
		- In general when plot command is called, ask if plotting should be done in current axis.
		- Add plot modifier from PyFDAP
		- Add additional plotting functions
		- Move all plotting functionalities in GUI to Plotting MB
		- In showObjPropsInBar, if array is longer than maxLength, write sth like "array/list of shape"
		- Get embryo/molecule compression saving running again.
		- Include 2D geometries, e.g. allow dim setting for custom geometries.
	- Path management. Check at installation if there is a installation of gmsh and fiji in executables folder. If so, set path to those.
	- Might be faster in analysis if we load all files at first, put them in common numpy array, apply all image manupulation steps
	  and finally grab ROIs via c=imgs[idxs,:]
	- pyfrp_gmsh_geometry.ruledSurface should store more than a single lineLoop but rather have a lineLoops list.
	- Need to add pyfrp_gmsh_geometry.planeSurface class.
	- Speed-up via trillionos
	- simulation.showIC via vtk
	- In GUI, when selecting fnFlatten, we need to make sure that we add slash

Additions:
	- Added scripting example.
	- Added pyfrp.modules.pyfrp_misc_module.addToPATHs .

Changes:

Bugfixes:
	- Fixed ValueError for None comparison in pyfrp_simulation.simulation.run.